Rename FileSystem to AbstractFileSystem
FileSystem is an ambient import which made the DX suffer when
refactoring code. It's also more obvious now that the FileSystem
interface is an interface that you need to implement, not something
concrete that you can use directly.
