+++ packages/theme-language-server-node/src/dependencies.spec.ts
@@ -8,7 +8,7 @@ import {
   filesForURI,
   findRootURI,
   getDefaultTranslationsFactory,
-  loadConfig
+  loadConfig,
 } from './dependencies';
 
 type Tree = {
