+++ packages/theme-language-server-node/src/index.ts
@@ -1,5 +1,5 @@
 import { startServer as startCoreServer } from '@shopify/theme-language-server-common';
-import { FileSystem, NodeFileSystem } from '@shopify/theme-check-node';
+import { AbstractFileSystem, NodeFileSystem } from '@shopify/theme-check-node';
 import { ThemeLiquidDocsManager } from '@shopify/theme-check-docs-updater';
 import { stdin, stdout } from 'node:process';
 import { createConnection } from 'vscode-languageserver/node';
@@ -12,7 +12,7 @@ import {
   loadConfig,
 } from './dependencies';
 
-export function startServer(fs: FileSystem = NodeFileSystem) {
+export function startServer(fs: AbstractFileSystem = NodeFileSystem) {
   const connection = createConnection(stdin, stdout);
   const log = (message: string) => console.error(message);
   const themeLiquidDocsManager = new ThemeLiquidDocsManager(log);
